Add back requirements.txt as generated by Pipfile for deployment
When this project is deployed from source via a requirements file,
not clear how to generate pip compatible setup on the fly as pipenv
seems to be required. The setup.py will fail without requirements.txt
already generated and it seems that having prerequisits for setup.py is
generally discouraged.  Add back requirements.txt as an interrum
workaround.